# web/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCA API
  version: 1.0.0
  description: |-
    Bank statements in, approved double-entry journal entries out.

    ### Two conventions worth knowing before you read anything else

    **Money is always a whole number of paise**, in a field named `*_paise`. Every amount also carries a `*_display` twin, already formatted with Indian digit grouping (`Rs 6,03,490.57`). Use the integer for arithmetic and the string for rendering, and never parse the string or format the integer yourself -- a JSON number with a decimal point becomes a float in a browser, which is exactly the rounding error the backend exists to avoid.

    **Nothing is final until a senior CA approves it.** Uploading a statement and classifying its rows produces *suggestions*. Only `POST /clients/{id}/approvals/` writes to the ledger, only a senior CA or firm admin may call it, and what it writes cannot afterwards be edited or deleted -- corrections are new entries that reverse and replace.

    ### Slow work

    Uploading a statement returns **202 Accepted** with a job. Poll `/jobs/{id}/` or subscribe to `/jobs/{id}/events/` for server-sent events.

    ### Errors

...
paths:
...
  /api/v1/jobs/{id}/:
    get:
      operationId: jobs_retrieve
      description: |-
        Progress on work the API did not block for.

        Every endpoint that starts slow work answers 202 with one of these. Poll it,
        or subscribe to ``events`` for the same thing pushed.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        required: true
      tags:
      - jobs
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: ''
...
components:
  schemas:
...
    Job:
      type: object
      description: |-
        What a 202 hands back, and what polling returns.

        ``result`` is whatever the work produced -- its shape depends on ``kind``
        and is documented on the endpoint that starts the job.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        kind:
          type: string
          readOnly: true
          description: e.g. statement.ingest
        status:
          allOf:
          - $ref: '#/components/schemas/JobStatusEnum'
          readOnly: true
        progress:
          type: integer
          readOnly: true
        message:
          type: string
          readOnly: true
        result:
          readOnly: true
        error:
          type: string
          readOnly: true
        error_code:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        started_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        finished_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - created_at
      - error
      - error_code
      - finished_at
      - id
      - kind
      - message
      - progress
      - result
      - started_at
      - status
    JobStatusEnum:
      enum:
      - PENDING
      - RUNNING
      - SUCCEEDED
      - FAILED
      type: string
      description: |-
        * `PENDING` - Queued
        * `RUNNING` - Running
        * `SUCCEEDED` - Succeeded
        * `FAILED` - Failed
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...